+++ internal/pkg/scaffold/cmd_test.go
@@ -165,11 +165,11 @@ func main() {
 
 // serveCRMetrics gets the Operator/CustomResource GVKs
 // and generates metrics based on those types.
-// It serves those metrics on "metricsHost/operatorMetricsPort".
+// It serves those metrics on "http://metricsHost:operatorMetricsPort".
 func serveCRMetrics(cfg *rest.Config) {
 	// Below returns filterted operator/CustomResource specific GVKs.
 	// For more control create below GVK list with your own custom logic.
-	filteredGVK, err := kubemetrics.GetGVKsFromAddToScheme(apis.AddToScheme)
+	filteredGVK, err := k8sutil.GetGVKsFromAddToScheme(apis.AddToScheme)
 	if err != nil {
 		log.Error(err, "Could not generate or serve Custom Resource metrics")
 		return
